Rebase project structure
