+++ src/main/java/com/carrepairshop/springframework/agh/controllers/MechanicController.java
@@ -1,7 +1,7 @@
 package com.carrepairshop.springframework.agh.controllers;
 
-import com.carrepairshop.springframework.agh.employees.AbstractEmployee;
-import com.carrepairshop.springframework.agh.services.EmployeeService;
+import com.carrepairshop.springframework.agh.domain.employees.AbstractEmployee;
+import com.carrepairshop.springframework.agh.service.EmployeeService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
