+++ src/main/java/com/carrepairshop/springframework/agh/dao/CarDao.java
@@ -1,6 +1,6 @@
 package com.carrepairshop.springframework.agh.dao;
 
-import com.carrepairshop.springframework.agh.car.Car;
+import com.carrepairshop.springframework.agh.domain.Car;
 
 import java.util.Collection;
 import java.util.Optional;
