+++ src/main/java/com/carrepairshop/springframework/agh/dao/HistoryDao.java
@@ -1,6 +1,6 @@
 package com.carrepairshop.springframework.agh.dao;
 
-import com.carrepairshop.springframework.agh.history.History;
+import com.carrepairshop.springframework.agh.domain.History;
 import java.util.Collection;
 
 public interface HistoryDao {
