+++ src/main/java/com/carrepairshop/springframework/agh/dao/OwnerDao.java
@@ -1,6 +1,6 @@
 package com.carrepairshop.springframework.agh.dao;
 
-import com.carrepairshop.springframework.agh.carOwner.Owner;
+import com.carrepairshop.springframework.agh.domain.Owner;
 
 import java.util.Collection;
