+++ src/main/java/com/carrepairshop/springframework/agh/dao/RepairDao.java
@@ -1,6 +1,6 @@
 package com.carrepairshop.springframework.agh.dao;
 
-import com.carrepairshop.springframework.agh.car.Repair;
+import com.carrepairshop.springframework.agh.domain.Repair;
 
 import java.util.Collection;
 import java.util.Optional;
