+++ src/main/java/com/carrepairshop/springframework/agh/dao/RoleDao.java
@@ -1,6 +1,6 @@
 package com.carrepairshop.springframework.agh.dao;
 
-import com.carrepairshop.springframework.agh.employees.Role;
+import com.carrepairshop.springframework.agh.domain.Role;
 
 import java.util.Collection;
 import java.util.Optional;
