+++ src/main/java/com/carrepairshop/springframework/agh/dao/impl/MySqlHistoryImpl.java
@@ -1,6 +1,7 @@
-package com.carrepairshop.springframework.agh.dao;
+package com.carrepairshop.springframework.agh.dao.impl;
 
-import com.carrepairshop.springframework.agh.history.History;
+import com.carrepairshop.springframework.agh.dao.HistoryDao;
+import com.carrepairshop.springframework.agh.domain.History;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
