+++ src/main/java/com/carrepairshop/springframework/agh/dao/impl/RoleImpl.java
@@ -1,6 +1,7 @@
-package com.carrepairshop.springframework.agh.dao;
+package com.carrepairshop.springframework.agh.dao.impl;
 
-import com.carrepairshop.springframework.agh.employees.Role;
+import com.carrepairshop.springframework.agh.dao.RoleDao;
+import com.carrepairshop.springframework.agh.domain.Role;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
