+++ src/main/java/com/carrepairshop/springframework/agh/domain/Car.java
@@ -1,4 +1,4 @@
-package com.carrepairshop.springframework.agh.car;
+package com.carrepairshop.springframework.agh.domain;
 
 
 import lombok.AllArgsConstructor;
