+++ src/main/java/com/carrepairshop/springframework/agh/domain/History.java
@@ -1,24 +1,23 @@
-package com.carrepairshop.springframework.agh.car;
+package com.carrepairshop.springframework.agh.domain;
 
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
-import lombok.NonNull;
 
 import javax.persistence.GeneratedValue;
 import javax.persistence.Id;
-import java.util.Date;
 
 @Data
 @AllArgsConstructor
 @NoArgsConstructor
-public class Appointment {
+public class History {
     @Id
     @GeneratedValue
     private Integer id;
-    @NonNull
-    private String nameUser;
-    private String numberCar;
-    private Date data;
+
+    private String date;
+    private String time;
     private String description;
+    private String type;
+
 }
