+++ src/main/java/com/carrepairshop/springframework/agh/domain/Role.java
@@ -1,4 +1,4 @@
-package com.carrepairshop.springframework.agh.employees;
+package com.carrepairshop.springframework.agh.domain;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
