+++ src/main/java/com/carrepairshop/springframework/agh/domain/calendars/WorkCalendar.java
@@ -1,4 +1,4 @@
-package com.carrepairshop.springframework.agh.calendars;
+package com.carrepairshop.springframework.agh.domain.calendars;
 
 import lombok.Data;
