+++ src/main/java/com/carrepairshop/springframework/agh/domain/employees/AbstractEmployee.java
@@ -1,4 +1,4 @@
-package com.carrepairshop.springframework.agh.employees;
+package com.carrepairshop.springframework.agh.domain.employees;
 
 import lombok.*;
