+++ src/main/java/com/carrepairshop/springframework/agh/domain/employees/Logistician.java
@@ -1,6 +1,6 @@
-package com.carrepairshop.springframework.agh.employees;
+package com.carrepairshop.springframework.agh.domain.employees;
 
-import com.carrepairshop.springframework.agh.calendars.WorkCalendar;
+import com.carrepairshop.springframework.agh.domain.calendars.WorkCalendar;
 import lombok.Builder;
 import lombok.Getter;
 import lombok.Setter;
