+++ src/main/java/com/carrepairshop/springframework/agh/service/CarService.java
@@ -1,6 +1,6 @@
-package com.carrepairshop.springframework.agh.services;
+package com.carrepairshop.springframework.agh.service;
 
-import com.carrepairshop.springframework.agh.car.Car;
+import com.carrepairshop.springframework.agh.domain.Car;
 import com.carrepairshop.springframework.agh.dao.CarDao;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Qualifier;
