+++ src/main/java/com/carrepairshop/springframework/agh/service/EmployeeService.java
@@ -1,7 +1,7 @@
-package com.carrepairshop.springframework.agh.services;
+package com.carrepairshop.springframework.agh.service;
 
 import com.carrepairshop.springframework.agh.dao.EmployeeDao;
-import com.carrepairshop.springframework.agh.employees.AbstractEmployee;
+import com.carrepairshop.springframework.agh.domain.employees.AbstractEmployee;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.stereotype.Service;
