+++ src/main/java/com/carrepairshop/springframework/agh/service/OwnerService.java
@@ -1,7 +1,7 @@
-package com.carrepairshop.springframework.agh.services;
+package com.carrepairshop.springframework.agh.service;
 
 
-import com.carrepairshop.springframework.agh.carOwner.Owner;
+import com.carrepairshop.springframework.agh.domain.Owner;
 import com.carrepairshop.springframework.agh.dao.OwnerDao;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Qualifier;
