+++ src/main/java/com/carrepairshop/springframework/agh/service/RepairService.java
@@ -1,6 +1,6 @@
-package com.carrepairshop.springframework.agh.services;
+package com.carrepairshop.springframework.agh.service;
 
-import com.carrepairshop.springframework.agh.car.Repair;
+import com.carrepairshop.springframework.agh.domain.Repair;
 import com.carrepairshop.springframework.agh.dao.RepairDao;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Qualifier;
